moves DateProvider
