+++ common/src/test/java/com/revenuecat/purchases/common/attributes/SubscriberAttributeTests.kt
@@ -1,7 +1,7 @@
 package com.revenuecat.purchases.common.attributes
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.revenuecat.purchases.common.DateProvider
+import com.revenuecat.purchases.utils.DateProvider
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatExceptionOfType
 import org.json.JSONException
