+++ feature/subscriber-attributes/build.gradle
@@ -2,6 +2,9 @@ apply from: "$rootProject.projectDir/library.gradle"
 
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
+    implementation project(":common")
+    implementation project(":utils")
+
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
     implementation 'androidx.core:core-ktx:1.3.1'
     implementation 'androidx.appcompat:appcompat:1.2.0'
